+++ src/main/kotlin/com/epam/brn/upload/csv/iterator/impl/SeriesMappingIteratorProvider.kt
@@ -1,29 +1,29 @@
 package com.epam.brn.upload.csv.iterator.impl
 
 import com.epam.brn.upload.csv.iterator.MappingIteratorProvider
-import com.epam.brn.upload.csv.record.SeriesRecord
+import com.epam.brn.upload.csv.record.SeriesGenericRecord
 import com.fasterxml.jackson.databind.MappingIterator
 import com.fasterxml.jackson.dataformat.csv.CsvMapper
 import com.fasterxml.jackson.dataformat.csv.CsvParser
 import java.io.InputStream
 import org.apache.commons.lang3.StringUtils
 
-class SeriesMappingIteratorProvider : MappingIteratorProvider<SeriesRecord> {
+class SeriesMappingIteratorProvider : MappingIteratorProvider<SeriesGenericRecord> {
 
-    override fun iterator(file: InputStream): MappingIterator<SeriesRecord> {
+    override fun iterator(file: InputStream): MappingIterator<SeriesGenericRecord> {
         val csvMapper = CsvMapper().apply {
             enable(CsvParser.Feature.TRIM_SPACES)
         }
 
         val csvSchema = csvMapper
-            .schemaFor(SeriesRecord::class.java)
+            .schemaFor(SeriesGenericRecord::class.java)
             .withColumnSeparator(',')
             .withLineSeparator(StringUtils.SPACE)
             .withColumnReordering(true)
             .withHeader()
 
         return csvMapper
-            .readerWithTypedSchemaFor(SeriesRecord::class.java)
+            .readerWithTypedSchemaFor(SeriesGenericRecord::class.java)
             .with(csvSchema)
             .readValues(file)
     }
