+++ src/main/kotlin/com/epam/brn/upload/csv/processor/SeriesGenericRecordProcessor.kt
@@ -1,18 +1,18 @@
-package com.epam.brn.upload.csv.converter.impl
+package com.epam.brn.upload.csv.processor
 
 import com.epam.brn.model.Series
 import com.epam.brn.service.ExerciseGroupsService
-import com.epam.brn.upload.csv.record.SeriesRecord
+import com.epam.brn.upload.csv.record.SeriesGenericRecord
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
 
 @Component
-class SeriesCsvConverter {
+class SeriesGenericRecordProcessor {
 
     @Autowired
     lateinit var exerciseGroupsService: ExerciseGroupsService
 
-    fun convert(source: SeriesRecord): Series {
+    fun convert(source: SeriesGenericRecord): Series {
         return Series(
             name = source.name,
             description = source.description,
