+++ src/test/kotlin/com/epam/brn/upload/csv/CsvParserTest.kt
@@ -4,8 +4,8 @@ import com.epam.brn.upload.csv.iterator.impl.GroupMappingIteratorProvider
 import com.epam.brn.upload.csv.iterator.impl.Series1TaskMappingIteratorProvider
 import com.epam.brn.upload.csv.iterator.impl.SeriesMappingIteratorProvider
 import com.epam.brn.upload.csv.record.GroupRecord
+import com.epam.brn.upload.csv.record.SeriesGenericRecord
 import com.epam.brn.upload.csv.record.SeriesOneTaskRecord
-import com.epam.brn.upload.csv.record.SeriesRecord
 import java.nio.charset.StandardCharsets
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -104,8 +104,8 @@ class CsvParserTest {
 
         assertThat(result).containsAll(
             listOf(
-                SeriesRecord(2, 1, "Распознование слов", "Распознование слов"),
-                SeriesRecord(2, 2, "Составление предложений", "Составление предложений")
+                SeriesGenericRecord(2, 1, "Распознование слов", "Распознование слов"),
+                SeriesGenericRecord(2, 2, "Составление предложений", "Составление предложений")
             )
         )
     }
